requirements.txt
fastapi[standard]==0.128.0
outline-python-api==0.4.0a0
python-dotenv==1.2.1
uvicorn==0.39.0
httpx==0.27.2
